requirements.txt
streamlit
spotipy
Pillow
langchain
langchain-google-genai
google-generativeai
faiss-cpu
tiktoken
# Add constraints for aiohttp to resolve typing error on Python 3.9
aiohttp==3.8.6
aiosignal==1.3.1
frozenlist==1.4.0